rrfactor: rename `definitions` to `application`
