+++ tests/application/course_schedule.rs
